upload original version of xPhoton and make it runable in  minimum modification
